clear bit

+++ 0x14-bit_manipulation/4-clear_bit.c
@@ -8,7 +8,7 @@
  */
 int clear_bit(unsigned long int *n, unsigned int index)
 {
-	unsigned int setBit = 1;
+	unsigned long int setBit = 1;
 
 	if (n == NULL)
 		return (-1);
